+++ packages/swirl-components/src/components/swirl-file-viewer/swirl-file-viewer.tsx
@@ -145,10 +145,13 @@ export class SwirlFileViewer {
                 ></swirl-file-viewer-audio>
               )}
 
+              {/* fallback */}
               {unsupportedType && (
-                <swirl-inline-error
-                  message={this.typeUnsupportedMessage}
-                ></swirl-inline-error>
+                <swirl-file-viewer-fallback
+                  disableDownload={this.disableDownload}
+                  file={this.file}
+                  fileName={this.fileName}
+                ></swirl-file-viewer-fallback>
               )}
             </div>
           )}